Use 32-bit float for fields that don't require double precision.

+++ src/point_one/fusion_engine/messages/solution.h
@@ -63,14 +63,14 @@ struct PoseMessage {
    * The X/Y/Z position standard deviation (in meters), resolved in the ECEF
    * frame.
    */
-  double position_std_dev_ecef_m[3] = {NAN, NAN, NAN};
+  float position_std_dev_ecef_m[3] = {NAN, NAN, NAN};
 
   /** The estimated aggregate 3D protection level (in meters). */
-  double aggregate_protection_level_m = NAN;
+  float aggregate_protection_level_m = NAN;
   /** The estimated 2D horizontal protection level (in meters). */
-  double horizontal_protection_level_m = NAN;
+  float horizontal_protection_level_m = NAN;
   /** The estimated vertical protection level (in meters). */
-  double vertical_protection_level_m = NAN;
+  float vertical_protection_level_m = NAN;
 };
 
 /**
@@ -101,13 +101,13 @@ struct GNSSInfoMessage {
   uint32_t reference_station_id = INVALID_REFERENCE_STATION;
 
   /** The geometric dilution of precision (GDOP). */
-  double gdop = NAN;
+  float gdop = NAN;
   /** The position dilution of precision (PDOP). */
-  double pdop = NAN;
+  float pdop = NAN;
   /** The horizontal dilution of precision (HDOP). */
-  double hdop = NAN;
+  float hdop = NAN;
   /** The vertical dilution of precision (VDOP). */
-  double vdop = NAN;
+  float vdop = NAN;
 
   /** The number of known satellites. */
   uint16_t num_satellites = 0;
@@ -138,10 +138,10 @@ struct SatelliteInfo {
   uint8_t reserved = 0;
 
   /** The azimuth of the satellite (in degrees). */
-  double azimuth_deg = NAN;
+  float azimuth_deg = NAN;
 
   /** The elevation of the satellite (in degrees). */
-  double elevation_deg = NAN;
+  float elevation_deg = NAN;
 };
 
 #pragma pack(pop)
